Enable third-party routing solutions
- consistently use uint16 for ports
- desired LRP routes become map of string to json.RawMessage

[#86337946]

+++ models/port_mapping.go
@@ -1,6 +1,6 @@
 package models
 
 type PortMapping struct {
-	ContainerPort uint32 `json:"container_port"`
-	HostPort      uint32 `json:"host_port,omitempty"`
+	ContainerPort uint16 `json:"container_port"`
+	HostPort      uint16 `json:"host_port,omitempty"`
 }
